criação de algumas classes e reafatoração

+++ src/Api.Invest/Model/Investimentos.cs
@@ -5,8 +5,8 @@ namespace Api.Invest.Model
     public class Investimentos
     {
         public string Nome { get; set; }
-        public double ValorTotal { get; set; }
-        public double ValorInvestido { get; set; }
+        public decimal ValorTotal { get; set; }
+        public decimal ValorInvestido { get; set; }
         public double ValorResgate { get; set; }
         public DateTime Vencimento { get; set; }
         public double IR { get; set; }
